Ditch specialist WavPlayer and make the AvPlayer work
Approach for player is to assume that audio is being played at 1:1 rate 
(nothing else is supported) and only make an effort to sync on player 
start, after that just feed the output stream as fast as it wants data. 
If the player position jumps then we detect this and resync.